requirements.txt
# before installing, you should install↓
# if you only use for training, please skip it.
# sudo apt-get install portaudio19-dev
# sudo apt install ffmpeg
# We recommend to use pytorch nightly version

# for training
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.0+cu111
-f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.9.0
hydra-core
omegaconf
joblib
matplotlib
tensorboard
pyopenjtalk
librosa
pysptk
pyworld
nnmnkwii
optuna
ttslearn
flake8
autopep8
mypy

# for others
jupyter
jupyterlab
tgt
types-PyYAML
pydub
pyaudio
sentence-transformers
sounddevice